1. Adjusted names of header files 2. Import of TimeFreq from LOPES-Tools code base
git-svn-id: http://usg.lofar.org/svn/code/trunk@252 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/CR-Pipeline/apps/simulator/mathclasses.cc
@@ -7,7 +7,7 @@ using namespace std;
 // Coordinate transformations have been tested and are ok at last!!!!
 
 //... Private header file with prototype for mathclasses
-#include "mathclasses.hh"
+#include "mathclasses.h"
 
 
 azelvec xyz2azel (double x, double y, double z) { 
@@ -58,8 +58,7 @@ double deg2rad(double deg){
 }
 
 double pi( void ) {
-
-  return acos(-1);
+  return acos(double(-1));
 }
 
 double c( void ) {
